+++ src/simulation/elements/SING.cpp
@@ -61,7 +61,7 @@ int SING_update(UPDATE_FUNC_ARGS)
 			}
 		}
 		spawncount = std::abs(parts[i].tmp);
-		spawncount = (spawncount>255) ? 3019 : std::pow((double)(spawncount/8), 2)*M_PI;
+		spawncount = (spawncount>255) ? 3019 : (int)(std::pow((double)(spawncount/8), 2)*M_PI);
 		for (int j = 0; j < spawncount; j++)
 		{
 			switch(rand()%3)
